complete calc, result, gold price, about pages

+++ app/src/main/java/com/midstatesrecycling/ktkalculator/model/KaratValue.kt
@@ -9,4 +9,4 @@ import kotlinx.parcelize.Parcelize
 @Keep
 data class KaratValue(
     @SerializedName("value")
-    val value: Float?) : Parcelable
+    val value: Double?) : Parcelable
